refactored and fixed uts
