+++ src/handlers/status-retriever-handler.ts
@@ -4,15 +4,10 @@ import logger from '../commons/logger';
 import { logAndPublishMetric } from '../commons/metrics';
 import { MetricNames, AISInterventionTypes } from '../data-types/constants';
 import { TransformedResponseFromDynamoDatabase } from '../data-types/interfaces';
-import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
-import { LOGS_PREFIX_SENSITIVE_INFO } from '../data-types/constants';
-import { QueryCommand, QueryCommandInput, QueryCommandOutput } from '@aws-sdk/client-dynamodb';
-import { TooManyRecordsError } from '../data-types/errors';
-import { unmarshall } from '@aws-sdk/util-dynamodb';
-import { NodeHttpHandler } from '@smithy/node-http-handler';
-import tracer from '../commons/tracer';
+import { DynamoDatabaseService } from '../services/dynamo-database-service';
 
 const appConfig = AppConfigService.getInstance();
+const dynamoDatabaseServiceInstance = new DynamoDatabaseService(appConfig.tableName);
 
 export const handle = async (event: APIGatewayEvent, context: Context): Promise<APIGatewayProxyResult> => {
   logger.addContext(context);
@@ -31,7 +26,7 @@ export const handle = async (event: APIGatewayEvent, context: Context): Promise<
   const historyQuery = event.queryStringParameters?.['history'];
 
   try {
-    const response = await queryRecordFromDynamoDatabase(userId);
+    const response = await dynamoDatabaseServiceInstance.queryRecordFromDynamoDatabase(userId);
     if (!response) {
       logger.warn('Query matched no records in DynamoDB.');
       logAndPublishMetric(MetricNames.ACCOUNT_NOT_FOUND);
@@ -112,38 +107,4 @@ function transformResponseFromDynamoDatabase(item: Record<string, any>): Transfo
     },
     auditLevel: String(item['auditLevel'] ?? 'standard'),
   });
-}
-
-async function queryRecordFromDynamoDatabase(userId: string) {
-  const dynamoClient = tracer.captureAWSv3Client(
-    new DynamoDBClient({
-      region: AppConfigService.getInstance().awsRegion,
-      maxAttempts: 2,
-      requestHandler: new NodeHttpHandler({ requestTimeout: 5000 }),
-    }),
-  );
-
-  logger.debug(`${LOGS_PREFIX_SENSITIVE_INFO} Attempting request to dynamo db, with ID : ${userId}`);
-  const parameters: QueryCommandInput = {
-    TableName: appConfig.tableName,
-    KeyConditionExpression: '#pk = :id_value',
-    ExpressionAttributeNames: { '#pk': 'pk' },
-    ExpressionAttributeValues: { ':id_value': { S: userId } },
-  };
-
-  const response: QueryCommandOutput = await dynamoClient.send(new QueryCommand(parameters));
-  if (!response.Items) {
-    const errorMessage = 'DynamoDB may have failed to query, returned a null response.';
-    logger.error(errorMessage);
-    logAndPublishMetric(MetricNames.DB_QUERY_ERROR_NO_RESPONSE);
-    throw new Error(errorMessage);
-  }
-
-  if (response.Items.length > 1) {
-    const errorMessage = 'DynamoDB returned more than one element.';
-    logger.error(errorMessage);
-    logAndPublishMetric(MetricNames.DB_QUERY_ERROR_TOO_MANY_ITEMS);
-    throw new TooManyRecordsError(errorMessage);
-  }
-  return response.Items[0] ? unmarshall(response.Items[0]) : undefined;
 }
